% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the first letter of each word}
\usage{
capitalize(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with each word capitalized.
}
\description{
Capitalize the first letter of each word
}
\examples{
capitalize("hello world")
# [1] "Hello World"
capitalize(c("apple pie", "banana split"))
}
